# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Find Your Car",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/fyc/clientCreds": {
            "get": {
                "description": "Get a list of all client credentials",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client Credentials"
                ],
                "summary": "Get all client credentials",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/pkg.ApiKey"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new client credential to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client Credentials"
                ],
                "summary": "Add a new client credential",
                "parameters": [
                    {
                        "description": "Client credential data",
                        "name": "clientCred",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg.ApiKey"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/pkg.ApiKey"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a client credential by ID",
                "tags": [
                    "Client Credentials"
                ],
                "summary": "Delete a client credential",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client credential deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/fyc/clientCreds/{id}": {
            "put": {
                "description": "Update an existing client credential by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client Credentials"
                ],
                "summary": "Update a client credential",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Updated client credential data",
                        "name": "clientCred",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg.ApiKey"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.ApiKey"
                        }
                    }
                }
            }
        },
        "/fyc/clientEnabled": {
            "get": {
                "description": "Get a list of enabled clients or a specific client by ID with optional extra data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client Credentials"
                ],
                "summary": "Get enabled clients or a specific client by clientID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of enabled clients or a single client",
                        "schema": {
                            "$ref": "#/definitions/pkg.ApiKey"
                        }
                    }
                }
            }
        },
        "/fyc/clientState": {
            "put": {
                "description": "Change the state of a Client (e.g., enabled/disabled) or retrieve a client by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client Credentials"
                ],
                "summary": "Change Client state or retrieve Client by ID",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Client State",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Client ID",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Number of rows affected by the state change",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "/fyc/clientsDeleted": {
            "get": {
                "description": "Get a list of deleted Client or a specific Client by ID with optional extra data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client Credentials"
                ],
                "summary": "Get deleted Clients or a specific Client by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of deleted Clients or a Client Client",
                        "schema": {
                            "$ref": "#/definitions/pkg.ApiKey"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "pkg.ApiKey": {
            "type": "object",
            "required": [
                "api_key",
                "client_id",
                "client_secret"
            ],
            "properties": {
                "api_key": {
                    "type": "string"
                },
                "client_id": {
                    "type": "string"
                },
                "client_name": {
                    "type": "string"
                },
                "client_secret": {
                    "type": "string"
                },
                "fuzzy": {
                    "type": "boolean",
                    "default": false
                },
                "id": {
                    "type": "integer"
                },
                "is_deleted": {
                    "type": "boolean",
                    "default": false
                },
                "is_enabled": {
                    "type": "boolean",
                    "default": false
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}